+++ cryptol-remote-api/python/cryptol/cryptoltypes.py
@@ -7,40 +7,110 @@ import BitVector #type: ignore
 from .bitvector import BV
 from .opaque import OpaqueValue
 
-from typing import Any, Dict, Iterable, List, NoReturn, Optional, TypeVar, Union
-
+import typing
+from typing import cast, Any, Dict, Iterable, List, NoReturn, Optional, TypeVar, Union
 from typing_extensions import Literal, Protocol
 
 A = TypeVar('A')
 
+def is_parenthesized(s : str) -> bool:
+    """Returns ``True`` iff the given string has balanced parentheses and is
+       enclosed in a matching pair of parentheses.
+       
+       :examples:
+       
+       >>> is_parenthesized(' ((a) b )')
+       True
+       >>> is_parenthesized('(a) (b)')
+       False
+       >>> is_parenthesized('(a')
+       False
+       """
+    seen_one, depth = False, 0
+    for c in s:
+        if depth > 0:
+            if c == '(': depth += 1
+            if c == ')': depth -= 1
+        else: # depth == 0
+            if c == '(':
+                if not seen_one: seen_one, depth = True, 1
+                # A new left paren after all parens have been closed means our
+                #  string is not enclosed in a matching pair of parentheses
+                else: return False
+            if c == ')':
+                # A right paren with no matching left means our string does
+                #  not have balanced parentheses
+                return False
+    # Return True if in the end all parentheses are balanced and we've seen at
+    #  least one matching pair
+    return seen_one and depth == 0
+
+def parenthesize(s : str) -> str:
+    """Encloses the given string ``s`` in parentheses if
+       ``is_parenthesized(s)`` is ``False``"""
+    return s if is_parenthesized(s) else f'({s})'
+
+
+JSON = Union[bool, int, float, str, Dict, typing.Tuple, List]
+
 class CryptolJSON(Protocol):
-    def __to_cryptol__(self, ty : CryptolType) -> Any: ...
+    def __to_cryptol__(self, ty : CryptolType) -> JSON: ...
+    def __to_cryptol_str__(self) -> str: ...
 
 class CryptolCode(metaclass=ABCMeta):
-    def __call__(self, other : CryptolJSON) -> CryptolCode:
-        return CryptolApplication(self, other)
+    def __call__(self, *others : CryptolJSON) -> CryptolCode:
+        if all(hasattr(other, '__to_cryptol__') for other in others):
+            return CryptolApplication(self, *others)
+        else:
+            raise ValueError("Argument to __call__ on CryptolCode is not CryptolJSON")
 
     @abstractmethod
-    def __to_cryptol__(self, ty : CryptolType) -> Any: ...
+    def __to_cryptol__(self, ty : CryptolType) -> JSON: ...
 
+    @abstractmethod
+    def __to_cryptol_str__(self) -> str: ...
+
+    def __str__(self) -> str:
+        return self.__to_cryptol_str__()
 
 class CryptolLiteral(CryptolCode):
     def __init__(self, code : str) -> None:
         self._code = code
 
-    def __to_cryptol__(self, ty : CryptolType) -> Any:
+    def __to_cryptol__(self, ty : CryptolType) -> JSON:
         return self._code
 
+    def __to_cryptol_str__(self) -> str:
+        return self._code
+
+    def __eq__(self, other : Any) -> bool:
+        return isinstance(other, CryptolLiteral) and self._code == other._code
+
+    def __repr__(self) -> str:
+        return f'CryptolLiteral({self._code!r})'
 
 class CryptolApplication(CryptolCode):
     def __init__(self, rator : CryptolJSON, *rands : CryptolJSON) -> None:
         self._rator = rator
         self._rands = rands
 
-    def __to_cryptol__(self, ty : CryptolType) -> Any:
+    def __to_cryptol__(self, ty : CryptolType) -> JSON:
         return {'expression': 'call',
                 'function': to_cryptol(self._rator),
                 'arguments': [to_cryptol(arg) for arg in self._rands]}
+
+    def __to_cryptol_str__(self) -> str:
+        if len(self._rands) == 0:
+            return self._rator.__to_cryptol_str__()
+        else:
+            return ' '.join(parenthesize(x.__to_cryptol_str__()) for x in [self._rator, *self._rands])
+
+    def __eq__(self, other : Any) -> bool:
+        return isinstance(other, CryptolApplication) and self._rator == other._rator and self._rands == other._rands
+
+    def __repr__(self) -> str:
+        return f'CryptolApplication({", ".join(repr(x) for x in [self._rator, *self._rands])})'
+
 
 class CryptolArrowKind:
     def __init__(self, dom : CryptolKind, ran : CryptolKind):
@@ -93,7 +163,7 @@ class Logic(UnaryProp):
         return f"Logic({self.subject!r})"
 
 
-def to_cryptol(val : Any, cryptol_type : Optional[CryptolType] = None) -> Any:
+def to_cryptol(val : Any, cryptol_type : Optional[CryptolType] = None) -> JSON:
     if cryptol_type is not None:
         return cryptol_type.from_python(val)
     else:
@@ -115,11 +185,12 @@ def is_plausible_json(val : Any) -> bool:
     return False
 
 class CryptolType:
-    def from_python(self, val : Any) -> Any:
+    def from_python(self, val : Any) -> JSON:
         if hasattr(val, '__to_cryptol__'):
             code = val.__to_cryptol__(self)
             if is_plausible_json(code):
-                return code
+                # the call to is_plausible_json ensures this cast is OK
+                return cast(JSON, code)
             else:
                 raise ValueError(f"Improbable JSON from __to_cryptol__: {val!r} gave {code!r}")
             # if isinstance(code, CryptolCode):
@@ -129,7 +200,7 @@ class CryptolType:
         else:
             return self.convert(val)
 
-    def convert(self, val : Any) -> Any:
+    def convert(self, val : Any) -> JSON:
         if isinstance(val, bool):
             return val
         elif isinstance(val, tuple) and val == ():