+++ cryptol-remote-api/python/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.12.2 -- YYYY-MM-DD
 
-* NEW CHANGELOG ENTRIES SINCE 2.12.0 GO HERE
+* Add an interface for Cryptol quasiquotation using an f-string-like syntax,
+  see `tests/cryptol/test_quoting` for some examples.
 
 ## 2.12.0 -- 2021-11-19